requirements.txt
python-dotenv
discord.py
pymongo
requests